# Specialize on the function and improve performance via StaticArrays