--- a/src/Gillespie.jl
+++ b/src/Gillespie.jl
@@ -4,6 +4,7 @@ using Distributions
 using DataFrames
 using QuadGK
 using Roots
+using StaticArrays
 
 export
     ssa,
@@ -34,7 +35,7 @@ There are several named arguments:
 - **thin**: (`Bool`) whether to thin jumps for Jensens method (default: `true`).
 
 "
-function ssa(x0::Vector{Int64},F::Base.Callable,nu::Matrix{Int64},parms::Vector{Float64},tf::Float64; algo=:gillespie, max_rate::Float64=0.0, thin::Bool=true)
+function ssa(x0::AbstractVector{Int64},F::Base.Callable,nu::AbstractMatrix{Int64},parms::AbstractVector{Float64},tf::Float64; algo=:gillespie, max_rate::Float64=0.0, thin::Bool=true)
   @assert algo in [:gillespie,:jensen,:tjm] "Available algorithms are :gillespie, :jensen, and :tjm"
   if algo == :gillespie
     return gillespie(x0,F,nu,parms,tf)