The DiffEq SSAStepper was achieving the same speed as Gillespie.jl, but that's odd because the DiffEq version is a sparse formulation that should have a lot better asymptotic scaling, but should have a little bit more overhead on the low end do to not inlining the affect! function. So I had to dig in and found a few ways to speed up Gillespie.jl. Basically, you can make this package quite a bit faster by making use of static arrays.

Benchmark:

```julia
using DiffEqBiological, OrdinaryDiffEq, Random, Statistics, Test
#using Plots

sir_model = @reaction_network rn begin
  0.1/1000, s + i --> 2i
  0.01, i --> r
end
sir_prob = DiscreteProblem([999,1,0],(0.0,250.0))
sir_jump_prob = JumpProblem(sir_prob,Direct(),sir_model)

sir_sol = solve(sir_jump_prob,SSAStepper())

#using Plots; plot(sir_sol)

Random.seed!(1234)
nums = Int[]
GC.gc()
@time for i in 1:100000
  sir_sol = solve(sir_jump_prob,SSAStepper())
  push!(nums,sir_sol[end][3])
end
println("Reaction DSL: $(mean(nums))")

using Gillespie

function F(x,parms)
  (S,I,R) = x
  (beta,gamma) = parms
  infection = beta*S*I
  recovery = gamma*I
  [infection,recovery]
end

x0 = [999,1,0]
nu = [[-1 1 0];[0 -1 1]]
parms = [0.1/1000.0,0.01]
tf = 250.0
Random.seed!(1234)

result = ssa(copy(x0),F,nu,parms,tf)
data = ssa_data(result)
push!(nums,data[!,:x3][end])

nums = Int[]
GC.gc()
@time for i in 1:100000
  result = ssa(copy(x0),F,nu,parms,tf)
  data = ssa_data(result)
  push!(nums,data[!,:x3][end])
end
println("Gillespie: $(mean(nums))")

using StaticArrays

function F2(x,parms)
  (S,I,R) = x
  (beta,gamma) = parms
  infection = beta*S*I
  recovery = gamma*I
  SA[infection,recovery]
end

x0 = SA[999,1,0]
nu = SA[-1 1 0
        0 -1 1]
parms = SA[0.1/1000.0,0.01]
tf = 250.0
Random.seed!(1234)

result = ssa(copy(x0),F2,nu,parms,tf)
data = ssa_data(result)
push!(nums,data[!,:x3][end])

nums = Int[]
GC.gc()
@time for i in 1:100000
  result = ssa(copy(x0),F2,nu,parms,tf)
  data = ssa_data(result)
  push!(nums,data[!,:x3][end])
end
println("Gillespie SA: $(mean(nums))")
```

Results:

```julia
25.371885 seconds (164.09 M allocations: 31.951 GiB, 9.02% gc time)
Reaction DSL: 725.7405
32.458954 seconds (326.63 M allocations: 34.374 GiB, 8.40% gc time)
Direct Jumps: 725.93672
25.931057 seconds (177.70 M allocations: 39.815 GiB, 12.76% gc time)
Gillespie: 725.53706
16.386420 seconds (15.54 M allocations: 25.301 GiB, 9.46% gc time)
Gillespie SA: 725.53706
```