@ChrisRackauckas 2020-05-11
@isaacsas note that with normal jumps the dynamic dispatch hit kicks in, and:

```julia
@inbounds rate = (u,p,t) -> (0.1/1000.0)*u[1]*u[2]
affect! = function (integrator)
  @inbounds integrator.u[1] -= 1
  @inbounds integrator.u[2] += 1
end
jump = ConstantRateJump(rate,affect!)

@inbounds rate = (u,p,t) -> 0.01u[2]
affect! = function (integrator)
  @inbounds integrator.u[2] -= 1
  @inbounds integrator.u[3] += 1
end
jump2 = ConstantRateJump(rate,affect!)


prob = DiscreteProblem([999,1,0],(0.0,250.0))
jump_prob = JumpProblem(prob,Direct(),jump,jump2)
sir_sol = solve(jump_prob,SSAStepper())

#using Plots; plot(sol)

nums = Int[]
GC.gc()
@time for i in 1:100000
  sol = solve(jump_prob,SSAStepper())
  push!(nums,sol[end][3])
end
println("Direct Jumps: $(mean(nums))")
```

```julia
31.162125 seconds (225.28 M allocations: 32.844 GiB, 7.31% gc time)
Direct Jumps: 724.74126
```

So this basically shows that it's the specialization on mass rate jumps that allows performance matching even with the sparse form.