@ChrisRackauckas 2020-05-11
With https://github.com/SciML/DiffEqJump.jl/pull/100 now the fastest one is DiffEq again:

```julia
using DiffEqBiological, OrdinaryDiffEq, Random, Statistics, Test
#using Plots

sir_model = @reaction_network rn begin
  0.1/1000, s + i --> 2i
  0.01, i --> r
end
sir_prob = DiscreteProblem(SA[999,1,0],(0.0,250.0))
sir_jump_prob = JumpProblem(sir_prob,Direct(),sir_model)

sir_sol = solve(sir_jump_prob,SSAStepper())

#using Plots; plot(sir_sol)

Random.seed!(1234)
nums = Int[]
GC.gc()
@time for i in 1:100000
  sir_sol = solve(sir_jump_prob,SSAStepper())
  push!(nums,sir_sol[end][3])
end
println("Reaction DSL: $(mean(nums))")
```

```julia
13.965592 seconds (1.38 M allocations: 29.879 GiB, 6.60% gc time)
Reaction DSL: 725.7405
```