@ChrisRackauckas 2020-05-11
The benchmarks were updated. Results on my computer:

|    Implementation                          | Time per simulation (ms) |
| -------------------------------------------| ------------------------ |
| Julia (Gillespie.jl)                       |          1.69            |
| Julia (Gillespie.jl, Static)               |          0.89            |
| Julia (DifferentialEquations.jl)           |          1.14            |
| Julia (DifferentialEquations.jl, Static)   |          0.72            |
| Julia (handcoded)                          |          0.49            |

Running the R next.