@ChrisRackauckas 2020-05-11
```
> print(paste("R:",sirsim/N*1000.0,"ms"))
[1] "R: 784.74 ms"
> print(paste("Rcpp:",sircsim/N*1000,"ms"))
[1] "Rcpp: 1.40000000000036 ms"
> print(paste("GillespieSSA:",sirssasim/N*1000,"ms"))
[1] "GillespieSSA: 463.609999999998 ms"
```

so the full results are:

|    Implementation                          | Time per simulation (ms) |
| -------------------------------------------| ------------------------ |
| R (GillespieSSA)                           |          463             |
| R (handcoded)                              |          785             |
| Rcpp (handcoded)                           |          1.40            |
| Julia (Gillespie.jl)                       |          1.69            |
| Julia (Gillespie.jl, Static)               |          0.89            |
| Julia (DifferentialEquations.jl)           |          1.14            |
| Julia (DifferentialEquations.jl, Static)   |          0.72            |
| Julia (handcoded)                          |          0.49            |