@sdwfrost 2020-05-11
Hi @ChrisRackauckas! I've meant to refactor some of this to avoid being all about Float64 arrays etc. for a while. In a recent example, I was still getting better performance with Gillespie.jl: see

[https://github.com/epirecipes/sir-julia/blob/master/markdown/jump_process_gillespie/jump_gillespie.md](https://github.com/epirecipes/sir-julia/blob/master/markdown/jump_process_gillespie/jump_gillespie.md)

vs.

[https://github.com/epirecipes/sir-julia/blob/master/markdown/jump_process/jump_process.md](https://github.com/epirecipes/sir-julia/blob/master/markdown/jump_process/jump_process.md)

Is this also a MassActionJump issue?