@isaacsas 2020-05-11
Chris can probably say more, but quickly looking at the latter there are two issues I can see. You should use `MassActionJump` for any normal mass action reactions. If your system has only jumps you should use the `SSAStepper` instead of `Functionmap`. The latter adds overhead by supporting many other features such as callbacks.