@isaacsas 2020-05-11
Also, DiffEqBiological will handle generating the fastest jump types that work for a given reaction for you, so you don't need to deal with specifying them individually.