@sdwfrost 2020-05-11
Thanks @isaacsas @ChrisRackauckas. I've updated my [epirecipes/sir-julia](https://github.com/epirecipes/sir-julia) repo accordingly. Off-topic, if you have any ideas on how to port [this example](https://github.com/epirecipes/sir-julia/blob/master/markdown/abm_vector/abm_vector.md) to DifferentialEquations, that would be great (I get a stack overflow if I just plug in FunctionMap).