@ChrisRackauckas 2020-05-12
Taking that to https://github.com/epirecipes/sir-julia/issues/4 and https://github.com/epirecipes/sir-julia/issues/5